vikingdjui/requirements.txt
fastapi
uvicorn
jinja2
pyserial
python-multipart
python-dotenv